chore: refactor scale_handler.go
