+++ pkg/scaling/scale_handler.go
@@ -85,6 +85,11 @@ func NewScaleHandler(client client.Client, scaleClient scale.ScalesGetter, recon
 	}
 }
 
+/// --------------------------------------------------------------------------- ///
+/// ----------            Scaling logic related methods               --------- ///
+/// --------------------------------------------------------------------------- ///
+
+// HandleScalableObject is the initial method when Scalable is created and it handles the main scaling logic
 func (h *scaleHandler) HandleScalableObject(ctx context.Context, scalableObject interface{}) error {
 	withTriggers, err := kedav1alpha1.AsDuckWithTriggers(scalableObject)
 	if err != nil {
@@ -122,6 +127,7 @@ func (h *scaleHandler) HandleScalableObject(ctx context.Context, scalableObject 
 	return nil
 }
 
+// DeleteScalableObject stops handling logic for input ScalableObject
 func (h *scaleHandler) DeleteScalableObject(ctx context.Context, scalableObject interface{}) error {
 	withTriggers, err := kedav1alpha1.AsDuckWithTriggers(scalableObject)
 	if err != nil {
@@ -149,7 +155,7 @@ func (h *scaleHandler) DeleteScalableObject(ctx context.Context, scalableObject 
 	return nil
 }
 
-// startScaleLoop blocks forever and checks the scaledObject based on its pollingInterval
+// startScaleLoop blocks forever and checks the scalableObject based on its pollingInterval
 func (h *scaleHandler) startScaleLoop(ctx context.Context, withTriggers *kedav1alpha1.WithTriggers, scalableObject interface{}, scalingMutex sync.Locker) {
 	logger := log.WithValues("type", withTriggers.Kind, "namespace", withTriggers.Namespace, "name", withTriggers.Name)
 
@@ -174,6 +180,84 @@ func (h *scaleHandler) startScaleLoop(ctx context.Context, withTriggers *kedav1a
 		}
 	}
 }
+
+// startPushScalers starts all push scalers defined in the input scalableOjbect
+func (h *scaleHandler) startPushScalers(ctx context.Context, withTriggers *kedav1alpha1.WithTriggers, scalableObject interface{}, scalingMutex sync.Locker) {
+	logger := log.WithValues("type", withTriggers.Kind, "namespace", withTriggers.Namespace, "name", withTriggers.Name)
+	cache, err := h.GetScalersCache(ctx, scalableObject)
+	if err != nil {
+		logger.Error(err, "Error getting scalers", "object", scalableObject)
+		return
+	}
+
+	for _, ps := range cache.GetPushScalers() {
+		go func(s scalers.PushScaler) {
+			activeCh := make(chan bool)
+			go s.Run(ctx, activeCh)
+			for {
+				select {
+				case <-ctx.Done():
+					return
+				case active := <-activeCh:
+					scalingMutex.Lock()
+					switch obj := scalableObject.(type) {
+					case *kedav1alpha1.ScaledObject:
+						h.scaleExecutor.RequestScale(ctx, obj, active, false)
+					case *kedav1alpha1.ScaledJob:
+						logger.Info("Warning: External Push Scaler does not support ScaledJob", "object", scalableObject)
+					}
+					scalingMutex.Unlock()
+				}
+			}
+		}(ps)
+	}
+}
+
+// checkScalers contains the main logic for the ScaleHandler scaling logic.
+// It'll check each trigger active status then call RequestScale
+func (h *scaleHandler) checkScalers(ctx context.Context, scalableObject interface{}, scalingMutex sync.Locker) {
+	scalingMutex.Lock()
+	defer scalingMutex.Unlock()
+	switch obj := scalableObject.(type) {
+	case *kedav1alpha1.ScaledObject:
+		err := h.client.Get(ctx, types.NamespacedName{Name: obj.Name, Namespace: obj.Namespace}, obj)
+		if err != nil {
+			log.Error(err, "error getting scaledObject", "object", scalableObject)
+			return
+		}
+		isActive, isError, metricsRecords, err := h.getScaledObjectState(ctx, obj)
+		if err != nil {
+			log.Error(err, "error getting state of scaledObject", "scaledObject.Namespace", obj.Namespace, "scaledObject.Name", obj.Name)
+			return
+		}
+
+		h.scaleExecutor.RequestScale(ctx, obj, isActive, isError)
+
+		if len(metricsRecords) > 0 {
+			log.V(1).Info("Storing metrics to cache", "scaledObject.Namespace", obj.Namespace, "scaledObject.Name", obj.Name, "metricsRecords", metricsRecords)
+			h.scaledObjectsMetricCache.StoreRecords(obj.GenerateIdentifier(), metricsRecords)
+		}
+	case *kedav1alpha1.ScaledJob:
+		cache, err := h.GetScalersCache(ctx, scalableObject)
+		if err != nil {
+			log.Error(err, "error getting scalers cache", "scaledJob.Namespace", obj.Namespace, "scaledJob.Name", obj.Name)
+			return
+		}
+
+		err = h.client.Get(ctx, types.NamespacedName{Name: obj.Name, Namespace: obj.Namespace}, obj)
+		if err != nil {
+			log.Error(err, "error getting scaledJob", "scaledJob.Namespace", obj.Namespace, "scaledJob.Name", obj.Name)
+			return
+		}
+
+		isActive, scaleTo, maxScale := cache.IsScaledJobActive(ctx, obj)
+		h.scaleExecutor.RequestJobScale(ctx, obj, isActive, scaleTo, maxScale)
+	}
+}
+
+/// --------------------------------------------------------------------------- ///
+/// ----------              ScalersCache related methods              --------- ///
+/// --------------------------------------------------------------------------- ///
 
 // GetScalersCache returns cache for input scalableObject, if the object is not found in the cache, it returns a new one
 // if the input object is ScaledObject, it also compares the Generation of the input of object with the one stored in the cache,
@@ -286,6 +370,7 @@ func (h *scaleHandler) performGetScalersCache(ctx context.Context, key string, s
 	return h.scalerCaches[key], nil
 }
 
+// ClearScalersCache invalidates chache for the input scalableObject
 func (h *scaleHandler) ClearScalersCache(ctx context.Context, scalableObject interface{}) error {
 	withTriggers, err := kedav1alpha1.AsDuckWithTriggers(scalableObject)
 	if err != nil {
@@ -307,71 +392,9 @@ func (h *scaleHandler) ClearScalersCache(ctx context.Context, scalableObject int
 	return nil
 }
 
-func (h *scaleHandler) startPushScalers(ctx context.Context, withTriggers *kedav1alpha1.WithTriggers, scalableObject interface{}, scalingMutex sync.Locker) {
-	logger := log.WithValues("type", withTriggers.Kind, "namespace", withTriggers.Namespace, "name", withTriggers.Name)
-	cache, err := h.GetScalersCache(ctx, scalableObject)
-	if err != nil {
-		logger.Error(err, "Error getting scalers", "object", scalableObject)
-		return
-	}
-
-	for _, ps := range cache.GetPushScalers() {
-		go func(s scalers.PushScaler) {
-			activeCh := make(chan bool)
-			go s.Run(ctx, activeCh)
-			for {
-				select {
-				case <-ctx.Done():
-					return
-				case active := <-activeCh:
-					scalingMutex.Lock()
-					switch obj := scalableObject.(type) {
-					case *kedav1alpha1.ScaledObject:
-						h.scaleExecutor.RequestScale(ctx, obj, active, false)
-					case *kedav1alpha1.ScaledJob:
-						logger.Info("Warning: External Push Scaler does not support ScaledJob", "object", scalableObject)
-					}
-					scalingMutex.Unlock()
-				}
-			}
-		}(ps)
-	}
-}
-
-// checkScalers contains the main logic for the ScaleHandler scaling logic.
-// It'll check each trigger active status then call RequestScale
-func (h *scaleHandler) checkScalers(ctx context.Context, scalableObject interface{}, scalingMutex sync.Locker) {
-	scalingMutex.Lock()
-	defer scalingMutex.Unlock()
-	switch obj := scalableObject.(type) {
-	case *kedav1alpha1.ScaledObject:
-		err := h.client.Get(ctx, types.NamespacedName{Name: obj.Name, Namespace: obj.Namespace}, obj)
-		if err != nil {
-			log.Error(err, "Error getting scaledObject", "object", scalableObject)
-			return
-		}
-		isActive, isError, metricsRecords := h.getScaledObjectState(ctx, obj)
-		h.scaleExecutor.RequestScale(ctx, obj, isActive, isError)
-		if len(metricsRecords) > 0 {
-			log.V(1).Info("Storing metrics to cache", "scaledObject.Namespace", obj.Namespace, "scaledObject.Name", obj.Name, "metricsRecords", metricsRecords)
-			h.scaledObjectsMetricCache.StoreRecords(obj.GenerateIdentifier(), metricsRecords)
-		}
-	case *kedav1alpha1.ScaledJob:
-		cache, err := h.GetScalersCache(ctx, scalableObject)
-		if err != nil {
-			log.Error(err, "Error getting scalers", "object", scalableObject)
-			return
-		}
-
-		err = h.client.Get(ctx, types.NamespacedName{Name: obj.Name, Namespace: obj.Namespace}, obj)
-		if err != nil {
-			log.Error(err, "Error getting scaledJob", "object", scalableObject)
-			return
-		}
-		isActive, scaleTo, maxScale := cache.IsScaledJobActive(ctx, obj)
-		h.scaleExecutor.RequestJobScale(ctx, obj, isActive, scaleTo, maxScale)
-	}
-}
+/// --------------------------------------------------------------------------- ///
+/// ----------             ScaledObject related methods               --------- ///
+/// --------------------------------------------------------------------------- ///
 
 // GetScaledObjectMetrics returns metrics for specified metric name for a ScaledObject identified by it's name and namespace.
 // The second return value are Prometheus metrics that needed to be exposed (used by DEPRECATED Prometheus Server on KEDA Metrics Server)
@@ -394,7 +417,7 @@ func (h *scaleHandler) GetScaledObjectMetrics(ctx context.Context, scaledObjectN
 	exportedPromMetrics.ScaledObjectErr = (err != nil)
 
 	if err != nil {
-		return nil, &exportedPromMetrics, fmt.Errorf("error when getting scalers %w", err)
+		return nil, &exportedPromMetrics, fmt.Errorf("error getting scalers %w", err)
 	}
 
 	var scaledObject *kedav1alpha1.ScaledObject
@@ -509,10 +532,12 @@ func (h *scaleHandler) GetScaledObjectMetrics(ctx context.Context, scaledObjectN
 	}, &exportedPromMetrics, nil
 }
 
-// getScaledObjectState returns whether the input ScaledObject is active as the first parameter,
-// the second parameter indicates whether there was any error during quering scalers
-// the third parameter returns map of metrics record - a metric value for each scaler and it's metric
-func (h *scaleHandler) getScaledObjectState(ctx context.Context, scaledObject *kedav1alpha1.ScaledObject) (bool, bool, map[string]metricscache.MetricsRecord) {
+// getScaledObjectState returns whether the input ScaledObject:
+// is active as the first return value,
+// the second return value indicates whether there was any error during quering scalers,
+// the third return value is a map of metrics record - a metric value for each scaler and it's metric
+// the fourth return value contains error if is not able access scalers cache
+func (h *scaleHandler) getScaledObjectState(ctx context.Context, scaledObject *kedav1alpha1.ScaledObject) (bool, bool, map[string]metricscache.MetricsRecord, error) {
 	logger := log.WithValues("scaledObject.Namespace", scaledObject.Namespace, "scaledObject.Name", scaledObject.Name)
 
 	isScaledObjectActive := false
@@ -521,6 +546,9 @@ func (h *scaleHandler) getScaledObjectState(ctx context.Context, scaledObject *k
 
 	cache, err := h.GetScalersCache(ctx, scaledObject)
 	prommetrics.RecordScaledObjectError(scaledObject.Namespace, scaledObject.Name, err)
+	if err != nil {
+		return false, true, map[string]metricscache.MetricsRecord{}, fmt.Errorf("error getting scalers cache %w", err)
+	}
 
 	// Let's collect status of all scalers, no matter if any scaler raises error or is active
 	scalers, scalerConfigs := cache.GetScalers()
@@ -596,201 +624,5 @@ func (h *scaleHandler) getScaledObjectState(ctx context.Context, scaledObject *k
 		logger.V(1).Info("scaler error encountered, clearing scaler cache")
 	}
 
-	return isScaledObjectActive, isScalerError, metricsRecord
-}
-
-// buildScalers returns list of Scalers for the specified triggers
-func (h *scaleHandler) buildScalers(ctx context.Context, withTriggers *kedav1alpha1.WithTriggers, podTemplateSpec *corev1.PodTemplateSpec, containerName string) ([]cache.ScalerBuilder, error) {
-	logger := log.WithValues("type", withTriggers.Kind, "namespace", withTriggers.Namespace, "name", withTriggers.Name)
-	var err error
-	resolvedEnv := make(map[string]string)
-	result := make([]cache.ScalerBuilder, 0, len(withTriggers.Spec.Triggers))
-
-	for i, t := range withTriggers.Spec.Triggers {
-		triggerIndex, trigger := i, t
-
-		factory := func() (scalers.Scaler, *scalers.ScalerConfig, error) {
-			if podTemplateSpec != nil {
-				resolvedEnv, err = resolver.ResolveContainerEnv(ctx, h.client, logger, &podTemplateSpec.Spec, containerName, withTriggers.Namespace, h.secretsLister)
-				if err != nil {
-					return nil, nil, fmt.Errorf("error resolving secrets for ScaleTarget: %w", err)
-				}
-			}
-			config := &scalers.ScalerConfig{
-				ScalableObjectName:      withTriggers.Name,
-				ScalableObjectNamespace: withTriggers.Namespace,
-				ScalableObjectType:      withTriggers.Kind,
-				TriggerName:             trigger.Name,
-				TriggerMetadata:         trigger.Metadata,
-				TriggerUseCachedMetrics: trigger.UseCachedMetrics,
-				ResolvedEnv:             resolvedEnv,
-				AuthParams:              make(map[string]string),
-				GlobalHTTPTimeout:       h.globalHTTPTimeout,
-				ScalerIndex:             triggerIndex,
-				MetricType:              trigger.MetricType,
-			}
-
-			authParams, podIdentity, err := resolver.ResolveAuthRefAndPodIdentity(ctx, h.client, logger, trigger.AuthenticationRef, podTemplateSpec, withTriggers.Namespace, h.secretsLister)
-			if err != nil {
-				return nil, nil, err
-			}
-			config.AuthParams = authParams
-			config.PodIdentity = podIdentity
-			scaler, err := buildScaler(ctx, h.client, trigger.Type, config)
-			return scaler, config, err
-		}
-
-		scaler, config, err := factory()
-		if err != nil {
-			h.recorder.Event(withTriggers, corev1.EventTypeWarning, eventreason.KEDAScalerFailed, err.Error())
-			logger.Error(err, "error resolving auth params", "scalerIndex", triggerIndex)
-			if scaler != nil {
-				scaler.Close(ctx)
-			}
-			for _, builder := range result {
-				builder.Scaler.Close(ctx)
-			}
-			return nil, err
-		}
-
-		result = append(result, cache.ScalerBuilder{
-			Scaler:       scaler,
-			ScalerConfig: *config,
-			Factory:      factory,
-		})
-	}
-
-	return result, nil
-}
-
-func buildScaler(ctx context.Context, client client.Client, triggerType string, config *scalers.ScalerConfig) (scalers.Scaler, error) {
-	// TRIGGERS-START
-	switch triggerType {
-	case "activemq":
-		return scalers.NewActiveMQScaler(config)
-	case "arangodb":
-		return scalers.NewArangoDBScaler(config)
-	case "artemis-queue":
-		return scalers.NewArtemisQueueScaler(config)
-	case "aws-cloudwatch":
-		return scalers.NewAwsCloudwatchScaler(config)
-	case "aws-dynamodb":
-		return scalers.NewAwsDynamoDBScaler(config)
-	case "aws-dynamodb-streams":
-		return scalers.NewAwsDynamoDBStreamsScaler(ctx, config)
-	case "aws-kinesis-stream":
-		return scalers.NewAwsKinesisStreamScaler(config)
-	case "aws-sqs-queue":
-		return scalers.NewAwsSqsQueueScaler(config)
-	case "azure-app-insights":
-		return scalers.NewAzureAppInsightsScaler(config)
-	case "azure-blob":
-		return scalers.NewAzureBlobScaler(config)
-	case "azure-data-explorer":
-		return scalers.NewAzureDataExplorerScaler(ctx, config)
-	case "azure-eventhub":
-		return scalers.NewAzureEventHubScaler(ctx, config)
-	case "azure-log-analytics":
-		return scalers.NewAzureLogAnalyticsScaler(config)
-	case "azure-monitor":
-		return scalers.NewAzureMonitorScaler(config)
-	case "azure-pipelines":
-		return scalers.NewAzurePipelinesScaler(ctx, config)
-	case "azure-queue":
-		return scalers.NewAzureQueueScaler(config)
-	case "azure-servicebus":
-		return scalers.NewAzureServiceBusScaler(ctx, config)
-	case "cassandra":
-		return scalers.NewCassandraScaler(config)
-	case "couchdb":
-		return scalers.NewCouchDBScaler(ctx, config)
-	case "cpu":
-		return scalers.NewCPUMemoryScaler(corev1.ResourceCPU, config)
-	case "cron":
-		return scalers.NewCronScaler(config)
-	case "datadog":
-		return scalers.NewDatadogScaler(ctx, config)
-	case "elasticsearch":
-		return scalers.NewElasticsearchScaler(config)
-	case "etcd":
-		return scalers.NewEtcdScaler(config)
-	case "external":
-		return scalers.NewExternalScaler(config)
-	// TODO: use other way for test.
-	case "external-mock":
-		return scalers.NewExternalMockScaler(config)
-	case "external-push":
-		return scalers.NewExternalPushScaler(config)
-	case "gcp-pubsub":
-		return scalers.NewPubSubScaler(config)
-	case "gcp-stackdriver":
-		return scalers.NewStackdriverScaler(ctx, config)
-	case "gcp-storage":
-		return scalers.NewGcsScaler(config)
-	case "graphite":
-		return scalers.NewGraphiteScaler(config)
-	case "huawei-cloudeye":
-		return scalers.NewHuaweiCloudeyeScaler(config)
-	case "ibmmq":
-		return scalers.NewIBMMQScaler(config)
-	case "influxdb":
-		return scalers.NewInfluxDBScaler(config)
-	case "kafka":
-		return scalers.NewKafkaScaler(config)
-	case "kubernetes-workload":
-		return scalers.NewKubernetesWorkloadScaler(client, config)
-	case "liiklus":
-		return scalers.NewLiiklusScaler(config)
-	case "loki":
-		return scalers.NewLokiScaler(config)
-	case "memory":
-		return scalers.NewCPUMemoryScaler(corev1.ResourceMemory, config)
-	case "metrics-api":
-		return scalers.NewMetricsAPIScaler(config)
-	case "mongodb":
-		return scalers.NewMongoDBScaler(ctx, config)
-	case "mssql":
-		return scalers.NewMSSQLScaler(config)
-	case "mysql":
-		return scalers.NewMySQLScaler(config)
-	case "nats-jetstream":
-		return scalers.NewNATSJetStreamScaler(config)
-	case "new-relic":
-		return scalers.NewNewRelicScaler(config)
-	case "openstack-metric":
-		return scalers.NewOpenstackMetricScaler(ctx, config)
-	case "openstack-swift":
-		return scalers.NewOpenstackSwiftScaler(ctx, config)
-	case "postgresql":
-		return scalers.NewPostgreSQLScaler(config)
-	case "predictkube":
-		return scalers.NewPredictKubeScaler(ctx, config)
-	case "prometheus":
-		return scalers.NewPrometheusScaler(config)
-	case "pulsar":
-		return scalers.NewPulsarScaler(config)
-	case "rabbitmq":
-		return scalers.NewRabbitMQScaler(config)
-	case "redis":
-		return scalers.NewRedisScaler(ctx, false, false, config)
-	case "redis-cluster":
-		return scalers.NewRedisScaler(ctx, true, false, config)
-	case "redis-cluster-streams":
-		return scalers.NewRedisStreamsScaler(ctx, true, false, config)
-	case "redis-sentinel":
-		return scalers.NewRedisScaler(ctx, false, true, config)
-	case "redis-sentinel-streams":
-		return scalers.NewRedisStreamsScaler(ctx, false, true, config)
-	case "redis-streams":
-		return scalers.NewRedisStreamsScaler(ctx, false, false, config)
-	case "selenium-grid":
-		return scalers.NewSeleniumGridScaler(config)
-	case "solace-event-queue":
-		return scalers.NewSolaceScaler(config)
-	case "stan":
-		return scalers.NewStanScaler(config)
-	default:
-		return nil, fmt.Errorf("no scaler found for type: %s", triggerType)
-	}
-	// TRIGGERS-END
+	return isScaledObjectActive, isScalerError, metricsRecord, nil
 }
